Various refactor tasks for server move
- Fix breaking typescript 2.0 changes in test files.
- Move 'test-utils' to lib directory and rename dir 'utils'.
- Convert logos to SVG in all files + adjust stylus to accommodate.
- Change the primary AIR color to the actual AIR color.
- Switch to docker-compose file composition - base file stored in 'lib'.
- Switch to named docker volumes using API version 2.
- Remove invalid npm scripts and replace with working scripts.
- Move 'scripts' directory to 'lib'
- './syncup.sh' command for new server will now be ran as an npm script
  in the form of 'npm run sync <args>'
