+++ aliemu/styles/colors.styl
@@ -3,7 +3,7 @@ $primary                = #00B092
 $secondary              = #3581B8
 
 // AIR Series Color Palette
-$air-primary               = #D0021B
+$air-primary               = #B91215
 $air-secondary             = #5B85AA
 $air-tertiary              = #EAD2AC
