+++ aliemu/templates/pages/air-pro.jade
@@ -1,11 +1,12 @@
 - get_header()
+| <?php $assets = get_stylesheet_directory_uri() . '/assets' ?>
 #main-content #[| <?php while (have_posts()) : the_post(); ?>]
     | <article id="post-<?php the_ID(); ?>" <?php post_class('category-air-pro'); ?>>
     .au-card
         .au-heading #[h1 Approved Instructional Resources Professional (AIR-Pro) series]
         .au-content
             .au-text #[div Welcome to the #[a(href="/about/#au-about-air-pro") Approved Instructional Resources Professional (AIR-Pro) series]! This series curates and grades open access blog posts and podcasts in the field of emergency medicine (EM) to identify high quality, social media-based, educational resources for senior EM residents. Unlike the AIR Series, this series targets more advanced practitioners and is no longer limited to content published within the previous 12 months. This series fulfills #[a(href="/about/#au-about-content") Individualized Interactive Instruction (III)], or asynchronous online learning, conference needs. Below are the AIR-Pro blocks along with their suggested number of III conference hour credit. If you are a Program Director and wish to track your residents’ progress, please request Educator Dashboard access.]
-            .au-image #[img(src="https://www.aliemu.com/wp-content/uploads/2016/06/airpro-badge.png", alt="AIR-Pro Badge")]
+            .au-image #[img(src="#{assets}/air-pro-logo.svg", alt="AIR-Pro Logo")]
     .au-links-row
         a(href="#year-2016") 2016
         a(href="#year-2015") 2015
