+++ aliemu/templates/pages/home.jade
@@ -24,8 +24,8 @@
                                 a(href="/user") My Profile
                                 a(href="/logout-2") Logout
                         .au-series-logos.au-alt-logos
-                            div #[a(href="/air") #[img(src="/wp-content/uploads/2016/06/air-button.png", alt="AIR Button")]]
-                            div #[a(href="/air-pro") #[img(src="/wp-content/uploads/2016/06/airpro-button.png", alt="AIR-Pro Button")]]
+                            div #[a(href="/air") #[img(src="#{assets}/air-logo.svg", alt="AIR Button")]]
+                            div #[a(href="/air-pro") #[img(src="#{assets}/air-pro-logo.svg", alt="AIR-Pro Button")]]
                             div #[a(href="/capsules") #[img(src="#{assets}/capsules-logo.svg", alt="Capsules Button")]]
         if !is_user_logged_in()
             #row-2
@@ -38,8 +38,8 @@
                     .au-info-card
                         .au-info-header #[h1 Click a series to learn more]
                         .au-series-logos
-                            div #[a(href="/air") #[img(src="/wp-content/uploads/2016/06/air-button.png", alt="AIR Button")]]
-                            div #[a(href="/air-pro") #[img(src="/wp-content/uploads/2016/06/airpro-button.png", alt="AIR-Pro Button")]]
+                            div #[a(href="/air") #[img(src="#{assets}/air-logo.svg", alt="AIR Button")]]
+                            div #[a(href="/air-pro") #[img(src="#{assets}/air-pro-logo.svg", alt="AIR-Pro Button")]]
                             div #[a(href="/capsules") #[img(src="#{assets}/capsules-logo.svg", alt="Capsules Button")]]
             #row-4
                 .au-container
