--- docker-compose.override.yml
@@ -1,26 +1,0 @@
-version: '2'
-services:
-    wordpress:
-        environment:
-            DB_USER: root
-            DB_PASS: root
-            LOCALHOST: 'true'
-            PLUGINS: >-
-                academic-bloggers-toolkit,
-                google-analytics-dashboard-for-wp,
-                query-monitor,
-                rest-api,
-                sendgrid-email-delivery-simplified,
-                tablepress,
-                tinymce-advanced,
-                ultimate-member,
-                wpdiscuz,
-                [Basic-Auth-master]https://github.com/WP-API/Basic-Auth/archive/master.zip,
-                [um-recaptcha]http://bit.ly/um-recaptcha,
-            SEARCH_REPLACE: https://www.aliemu.com,http://localhost:8080
-            WP_DEBUG: 'true'
-            WP_DEBUG_LOG: 'true'
-            WP_DEBUG_DISPLAY: 'false'
-    db:
-        environment:
-            MYSQL_ROOT_PASSWORD: root
