+++ docker-compose.staging.yml
@@ -1,7 +1,20 @@
 version: '2'
 services:
     wordpress:
+        extends:
+            file: lib/compose.yml
+            service: wordpress_prod
+        links:
+            - db
         environment:
             SEARCH_REPLACE: https://www.aliemu.com,http://45.55.237.250:8080
             LOCALHOST: 'true'
-            DEBUG: 'true'
+            DB_PASS: root
+    db:
+        extends:
+            file: lib/compose.yml
+            service: db_prod
+        environment:
+            MYSQL_ROOT_PASSWORD: root
+volumes:
+    data: {}
