+++ docker-compose.yml
@@ -1,38 +1,18 @@
 version: '2'
 services:
     wordpress:
-        image: dsifford/wordpress:nginx
+        extends:
+            file: lib/compose.yml
+            service: wordpress
         links:
             - db
-        ports:
-            - 22222:22222
-            - 8080:80
-            - 443:443
-        volumes:
-            # Database
-            - ./data:/data
-            # Plugins
-            - ./wp-content/plugins/learndash-propanel:/var/www/aliemu.com/htdocs/wp-content/plugins/learndash-propanel
-            - ./wp-content/plugins/sfwd-lms:/var/www/aliemu.com/htdocs/wp-content/plugins/sfwd-lms
-            # Themes
-            - ./dist/Divi-child:/var/www/aliemu.com/htdocs/wp-content/themes/Divi-child
-            - ./wp-content/themes/Divi:/var/www/aliemu.com/htdocs/wp-content/themes/Divi
-            # Uploads
-            - ./wp-content/uploads:/var/www/aliemu.com/htdocs/wp-content/uploads
         environment:
-            DB_NAME: wordpress
-            SITE_NAME: aliemu.com
-            PHP_VERSION: 7.0
-            WP_DEBUG: 'true'
-            WP_DEBUG_LOG: 'true'
-            WP_DEBUG_DISPLAY: 'false'
+            LOCALHOST: 'true'
+            SEARCH_REPLACE: https://www.aliemu.com,http://localhost:8080
+            WP_DEBUG_DISPLAY: 'true'
     db:
-        image: mariadb:10
-        ports:
-            - 3306:3306
-        volumes_from:
-            - data
-    data:
-        image: busybox
-        volumes:
-            - /var/lib/mysql
+        extends:
+            file: lib/compose.yml
+            service: db
+volumes:
+    data: {}
