+++ lib/compose.yml
@@ -1,0 +1,59 @@
+version: '2'
+services:
+    wordpress:
+        image: dsifford/wordpress:nginx
+        ports:
+            - "22222:22222"
+            - "8080:80"
+            - "443:443"
+        volumes:
+            # Database
+            - ../data:/data
+            # Plugins
+            - ../wp-content/plugins/learndash-propanel:/var/www/aliemu.com/htdocs/wp-content/plugins/learndash-propanel
+            - ../wp-content/plugins/sfwd-lms:/var/www/aliemu.com/htdocs/wp-content/plugins/sfwd-lms
+            # Themes
+            - ../dist/Divi-child:/var/www/aliemu.com/htdocs/wp-content/themes/Divi-child
+            - ../wp-content/themes/Divi:/var/www/aliemu.com/htdocs/wp-content/themes/Divi
+            # Uploads
+            - ../wp-content/uploads:/var/www/aliemu.com/htdocs/wp-content/uploads
+        env_file: ../.env
+        environment:
+            DB_NAME: wordpress
+            DB_PASS: root
+            DB_USER: root
+            PHP_VERSION: 7.0
+            PLUGINS: >-
+                academic-bloggers-toolkit,
+                google-analytics-dashboard-for-wp,
+                query-monitor,
+                rest-api,
+                sendgrid-email-delivery-simplified,
+                tablepress,
+                tinymce-advanced,
+                ultimate-member,
+                wpdiscuz,
+                [Basic-Auth-master]https://github.com/WP-API/Basic-Auth/archive/master.zip,
+                [um-recaptcha]http://bit.ly/um-recaptcha,
+            SITE_NAME: aliemu.com
+            WP_DEBUG_DISPLAY: 'false'
+            WP_DEBUG_LOG: 'true'
+            WP_DEBUG: 'true'
+    db:
+        image: mariadb:10
+        expose:
+            - "3306"
+        environment:
+            MYSQL_ROOT_PASSWORD: root
+        volumes:
+            - data:/var/lib/mysql
+    wordpress_prod:
+        extends: wordpress
+        restart: on-failure:10
+        environment:
+            DB_PASS: $DB_PASS
+    db_prod:
+        extends: db
+        restart: on-failure:10
+        environment:
+            MYSQL_ROOT_PASSWORD: $DB_PASS
