+++ lib/scripts/backup
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 SCRIPTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-ROOTDIR=$(cd $SCRIPTDIR && cd ../ && pwd || exit)
+ROOTDIR=$(cd $SCRIPTDIR && cd ../../ && pwd || exit)
 FILENAME=$(date +%Y-%m-%d_%H-%M_aliemu)
 
 cd $ROOTDIR/data || exit 1
