+++ lib/scripts/deploy
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 SCRIPTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-ROOTDIR=$(cd $SCRIPTDIR && cd ../ && pwd || exit)
+ROOTDIR=$(cd $SCRIPTDIR && cd ../../ && pwd || exit)
 
 # Pretty colors
 ORANGE='\033[0;33m'
