+++ lib/scripts/deploy-new
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 SCRIPTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-ROOTDIR=$(cd $SCRIPTDIR && cd ../ && pwd || exit)
+ROOTDIR=$(cd $SCRIPTDIR && cd ../../ && pwd || exit)
 machine_name=aliemu
 
 # shellcheck source=./helpers
