+++ lib/typings/ALiEMU.d.ts
